requirements.txt
tkinter=0.1.0
av=10.0.0
pillow=9.3.0
numpy=1.23.5
imageio=2.22.4
